server/pyproject.toml
[tool.poetry]
name = "diyivi"
version = "0.1.0"
description = "Backend for DIYivi"
authors = ["Dirk Doesburg <dirk.doesburg@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.0"}
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.4"
pyjwt = {extras = ["cryptography"], version = "^2.8.0"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pre-commit = "^3.7.1"
mypy = "^1.10.1"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.lint.isort]
known-first-party = ["app"]